% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa4ss.R
\docType{package}
\name{sa4all}
\alias{sa4all}
\title{sa4all: Generate stock assessment pdfs from Stock Synthesis (SS)}
\description{
The \pkg{sa4all} is designed to facilitate the easy construction of
stock assessment documents for species that were modelled with
Stock Synthesis (SS), a statistical catch-at-age assessment framework.
}
